requirements.txt
openai
python-dotenv
openai-whisper
tiktoken

